move AnalyzeToolbar to separate file
